#1 fix script error handling by using shell sets; replace if brackets
The shell script did not react in any way towards any occurred errors.
This commit adds the usual bash handling: If any command or pipe fails
the whole script returns with an exit code != zero, indicating there was
an error.

Furthermore, single brackets in bash scripts are discouraged. For more
if-test safety one occurrence was replaced with double brackets which
is supposed to have
[less surprises](https://stackoverflow.com/questions/669452/).